Add files via upload
Both the Financial Dashboard and Pivot_Table files are part of the excel assignment in TASK II.
The data_file and the Test_data_file.csv are the files for Python or R programming assignments under Machine Learning.
The Work in Progress. Rmd file is the R markdown file with codes and interpretations.